+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import About from '~/sections/About/About';
-import Contact from '~/sections/Contact';
-import Hero from '~/sections/Hero';
+import Contact from '~/sections/Contact/Contact';
+import Hero from '~/sections/Hero/Hero';
 import Portfolio from '~/sections/Portfolio/Portfolio';
 
 function HomePage() {
