+++ src/sections/About/AboutBody.tsx
@@ -73,7 +73,7 @@ const AboutBody: FC<AboutBodyProps> = ({ githubProfileData }) => {
   }, [isImageInView]);
 
   return (
-    <div ref={headerRef} className="about-body">
+    <div ref={headerRef} className="about-body flex flex-col items-center">
       <Heading
         as="h1"
         className="section-header flex flex-col items-center justify-center text-center"
@@ -82,14 +82,14 @@ const AboutBody: FC<AboutBodyProps> = ({ githubProfileData }) => {
         <Underline className="header-text min-w-[250px] max-w-xl bg-brandLight-500 px-4 dark:bg-brandDark-500" />
       </Heading>
 
-      <p className="flavor-text mb-10 text-center text-lg">
+      <p className="flavor-text mb-10 hyphens-auto break-words text-center text-base sm:text-lg">
         {githubProfileData?.bio ?? '<-- Skills to pay the bills -->'}
       </p>
       <div
         ref={imageRef}
         className="grid grid-cols-1 gap-8 px-4 sm:grid-cols-2 md:gap-6"
       >
-        <div className="image-wrapper relative aspect-square w-full max-w-md origin-center overflow-clip rounded-2xl">
+        <div className="image-wrapper relative mx-auto aspect-square w-full max-w-md origin-center overflow-clip rounded-2xl">
           <Image
             src={githubProfileData?.avatar_url ?? '/default-profile-pic.jpeg'}
             fill={true}
